+++ src/Controller/Authentification/RegistrationController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Authentification;
 
 use App\Entity\User;
 use App\Form\RegistrationFormType;
